Update Feb 22 19 Tab Creator
update @ 7:50 pm. Unable to get input (where notated) to correctly jump back to the top of the loop and skip writing any code to the file. Switching around the clause to write to a file, and to continue, will cause the continue to happen to matter what input the user enters and will never write to the file. First statement after cin >> choice1; is always executed.